refactor: Separate server and client components for course edit page.
